+++ server/routes/session.js
@@ -1,55 +1,16 @@
 const express = require("express");
-const { check } = require("express-validator");
-
-const { setTokenCookie, restoreUser } = require("../utils/auth");
-const { handleValidationErrors } = require("../utils/validation");
-const User = require("../models/User");
-
 const router = express.Router();
 
-const validateLogin = [
-  check("credential")
-    .exists({ checkFalsy: true })
-    .notEmpty()
-    .withMessage("Please provide a valid email or username."),
-  check("password")
-    .exists({ checkFalsy: true })
-    .withMessage("Please provide a password."),
-  handleValidationErrors,
-];
+const { validateLogin } = require("../utils/validation");
+const {
+  loginUser,
+  logoutUser,
+  restoreXsrfToken,
+} = require("../controllers/sessionController");
 
-router.use(restoreUser);
-
+router.post("/", validateLogin, loginUser);
+router.delete("/", logoutUser);
 // should not be available in production
-router.get("/csrf/restore", (req, res) => {
-  const csrfToken = req.csrfToken();
-  res.cookie("XSRF-TOKEN", csrfToken);
-  res.status(200).json({
-    "XSRF-Token": csrfToken,
-  });
-});
-
-router.post("/", validateLogin, async (req, res, next) => {
-  const { credential, password } = req.body;
-
-  const user = await User.loginUser(credential, password);
-
-  if (!user) {
-    const err = new Error("Login failed. Invalid credentials.");
-    err.status = 401;
-    return next(err);
-  }
-
-  await setTokenCookie(res, user);
-
-  return res.json({
-    user: user,
-  });
-});
-
-router.delete("/", (_req, res) => {
-  res.clearCookie("token");
-  return res.json({ message: "Logout successful." });
-});
+router.get("/csrf/restore", restoreXsrfToken);
 
 module.exports = router;
