+++ server/routes/users.js
@@ -1,8 +1,7 @@
 var express = require("express");
 var router = express.Router();
-const { check } = require("express-validator");
 
-const { handleValidationErrors } = require("../utils/validation");
+const { validateSignup } = require("../utils/validation");
 const {
   getAllUsers,
   getSingleUser,
@@ -13,23 +12,6 @@ const {
   deleteUser,
   updateUser,
 } = require("../controllers/usersController");
-
-const validateSignup = [
-  check("email")
-    .exists({ checkFalsy: true })
-    .isEmail()
-    .withMessage("Please provide a valid email."),
-  check("username")
-    .exists({ checkFalsy: true })
-    .isLength({ min: 4 })
-    .withMessage("Please provide a username with at least 4 characters."),
-  check("username").not().isEmail().withMessage("Username cannot be an email."),
-  check("password")
-    .exists({ checkFalsy: true })
-    .isLength({ min: 6 })
-    .withMessage("Password must be 6 characters or more."),
-  handleValidationErrors,
-];
 
 router.get("/", getAllUsers);
 router.post("/", validateSignup, createUser);
